chore: cleanup comments and reuse create_test_register_store
