+++ Cargo.toml
@@ -77,6 +77,7 @@ exponential-backoff = "1.0.0"
 async-recursion = "0.3.2"
 sled = "0.34.6"
 sysinfo = "0.19.0"
+tempfile = "3.2.0"
 
   [dependencies.ed25519-dalek]
   version = "1.0.0"
@@ -124,7 +125,6 @@ assert_matches = "1.3"
 proptest = "0.10.1"
 rand_xorshift = "~0.2.0"
 structopt = "~0.3.17"
-tempfile = "3.2.0"
 yansi = "~0.5.0"
 criterion = { version="0.3", features=["async_tokio"]}
